normed & fixed

+++ libft/srcs/ft_put/ft_putwchar_fd.c
@@ -6,7 +6,7 @@
 /*   By: ttroll <ttroll@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/03 13:15:55 by ttroll            #+#    #+#             */
-/*   Updated: 2019/04/22 18:59:05 by ttroll           ###   ########.fr       */
+/*   Updated: 2019/09/11 15:25:53 by ttroll           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,9 +41,9 @@ static int				err_handle(wint_t c, int fd)
 
 int						ft_putwchar_fd(wint_t c, int fd)
 {
-	size_t				size;
-	unsigned char	*arr;
-	int				i;
+	int					size;
+	unsigned char		*arr;
+	int					i;
 
 	i = 0;
 	size = (size_t)get_wchar_size(c);
